+++ graphql_client_cli/src/main.rs
@@ -1,7 +1,5 @@
 extern crate failure;
 extern crate reqwest;
-
-#[macro_use]
 extern crate structopt;
 #[macro_use]
 extern crate graphql_client;
@@ -10,6 +8,8 @@ extern crate graphql_client_codegen;
 extern crate serde_derive;
 extern crate serde;
 extern crate serde_json;
+#[macro_use]
+extern crate syn;
 
 #[cfg(feature = "rustfmt")]
 extern crate rustfmt_nightly as rustfmt;
@@ -60,6 +60,10 @@ enum Cli {
         /// Formating feature is disabled as default installation.
         #[structopt(long = "no-formatting")]
         no_formatting: bool,
+        /// You can choose module and target struct visibility from pub and private.
+        /// Default value is pub.
+        #[structopt(short = "m", long = "module_visibility")]
+        module_visibility: Option<String>,
         #[structopt(parse(from_os_str))]
         output: PathBuf,
     },
@@ -81,6 +85,7 @@ fn main() -> Result<(), failure::Error> {
             additional_derives,
             deprecation_strategy,
             no_formatting,
+            module_visibility,
             output,
         } => generate::generate_code(
             query_path,
@@ -90,6 +95,7 @@ fn main() -> Result<(), failure::Error> {
             additional_derives,
             &deprecation_strategy,
             no_formatting,
+            &module_visibility,
             &output,
         ),
     }