+++ graphql_client_codegen/src/interfaces.rs
@@ -2,15 +2,17 @@ use failure;
 use objects::GqlObjectField;
 use proc_macro2::{Ident, Span, TokenStream};
 use query::QueryContext;
-use selection::{Selection, SelectionItem};
+use selection::{Selection, SelectionField, SelectionFragmentSpread, SelectionItem};
 use shared::*;
 use std::borrow::Cow;
 use std::cell::Cell;
 use std::collections::HashSet;
 use unions::union_variants;
 
+/// Represents an Interface type extracted from the schema.
 #[derive(Debug, Clone, PartialEq)]
 pub struct GqlInterface {
+    /// The documentation for the interface. Extracted from the schema.
     pub description: Option<String>,
     /// The set of object types implementing this interface.
     pub implemented_by: HashSet<String>,
@@ -23,7 +25,7 @@ pub struct GqlInterface {
 
 impl GqlInterface {
     /// filters the selection to keep only the fields that refer to the interface's own.
-    fn object_selection(&self, selection: &Selection) -> Selection {
+    fn object_selection(&self, selection: &Selection, query_context: &QueryContext) -> Selection {
         Selection(
             selection
                 .0
@@ -31,8 +33,43 @@ impl GqlInterface {
                 // Only keep what we can handle
                 .filter(|f| match f {
                     SelectionItem::Field(f) => f.name != "__typename",
-                    SelectionItem::FragmentSpread(_) => true,
+                    SelectionItem::FragmentSpread(SelectionFragmentSpread { fragment_name }) => {
+                        // only if the fragment refers to the interface’s own fields (to take into account type-refining fragments)
+                        let fragment = query_context
+                            .fragments
+                            .get(fragment_name)
+                            .ok_or_else(|| format_err!("Unknown fragment: {}", &fragment_name))
+                            // TODO: fix this
+                            .unwrap();
+
+                        fragment.on == self.name
+                    }
                     SelectionItem::InlineFragment(_) => false,
+                }).map(|a| (*a).clone())
+                .collect(),
+        )
+    }
+
+    fn union_selection(&self, selection: &Selection, query_context: &QueryContext) -> Selection {
+        Selection(
+            selection
+                .0
+                .iter()
+                // Only keep what we can handle
+                .filter(|f| match f {
+                    SelectionItem::InlineFragment(_) => true,
+                    SelectionItem::FragmentSpread(SelectionFragmentSpread { fragment_name }) => {
+                        let fragment = query_context
+                            .fragments
+                            .get(fragment_name)
+                            .ok_or_else(|| format_err!("Unknown fragment: {}", &fragment_name))
+                            // TODO: fix this
+                            .unwrap();
+
+                        // only the fragments _not_ on the interface
+                        fragment.on != self.name
+                    }
+                    SelectionItem::Field(SelectionField { name, .. }) => name == "__typename",
                 }).map(|a| (*a).clone())
                 .collect(),
         )
@@ -59,7 +96,7 @@ impl GqlInterface {
         ::shared::field_impls_for_selection(
             &self.fields,
             context,
-            &self.object_selection(selection),
+            &self.object_selection(selection, context),
             prefix,
         )
     }
@@ -75,7 +112,7 @@ impl GqlInterface {
             &self.name,
             &self.fields,
             context,
-            &self.object_selection(selection),
+            &self.object_selection(selection, context),
             prefix,
         )
     }
@@ -90,26 +127,21 @@ impl GqlInterface {
         let name = Ident::new(&prefix, Span::call_site());
         let derives = query_context.response_derives();
 
-        selection
-            .extract_typename()
-            .ok_or_else(|| format_err!("Missing __typename in selection for {}", prefix))?;
-
-        let union_selection = Selection(
-            selection
-                .0
-                .iter()
-                // Only keep what we can handle
-                .filter(|f| match f {
-                    SelectionItem::InlineFragment(_) => true,
-                    SelectionItem::Field(_) | SelectionItem::FragmentSpread(_) => false,
-                }).map(|a| (*a).clone())
-                .collect(),
-        );
+        selection.extract_typename().ok_or_else(|| {
+            format_err!(
+                "Missing __typename in selection for the {} interface (type: {})",
+                prefix,
+                self.name
+            )
+        })?;
 
         let object_fields =
             self.response_fields_for_selection(query_context, &selection, prefix)?;
 
         let object_children = self.field_impls_for_selection(query_context, &selection, prefix)?;
+
+        let union_selection = self.union_selection(&selection, &query_context);
+
         let (mut union_variants, union_children, used_variants) =
             union_variants(&union_selection, query_context, prefix)?;
 
@@ -152,5 +184,62 @@ impl GqlInterface {
                 #last_object_field
             }
         })
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    // to be improved
+    #[test]
+    fn union_selection_works() {
+        let iface = GqlInterface {
+            description: None,
+            implemented_by: HashSet::new(),
+            name: "MyInterface".into(),
+            fields: vec![],
+            is_required: Cell::new(true),
+        };
+
+        let context = QueryContext::new_empty();
+
+        let typename_field = ::selection::SelectionItem::Field(::selection::SelectionField {
+            alias: None,
+            name: "__typename".to_string(),
+            fields: Selection(vec![]),
+        });
+        let selection = Selection(vec![typename_field.clone()]);
+
+        assert_eq!(
+            iface.union_selection(&selection, &context),
+            Selection(vec![typename_field])
+        );
+    }
+
+    // to be improved
+    #[test]
+    fn object_selection_works() {
+        let iface = GqlInterface {
+            description: None,
+            implemented_by: HashSet::new(),
+            name: "MyInterface".into(),
+            fields: vec![],
+            is_required: Cell::new(true),
+        };
+
+        let context = QueryContext::new_empty();
+
+        let typename_field = ::selection::SelectionItem::Field(::selection::SelectionField {
+            alias: None,
+            name: "__typename".to_string(),
+            fields: Selection(vec![]),
+        });
+        let selection = Selection(vec![typename_field]);
+
+        assert_eq!(
+            iface.object_selection(&selection, &context),
+            Selection(vec![typename_field])
+        );
     }
 }