+++ graphql_client_codegen/src/unions.rs
@@ -2,7 +2,7 @@ use constants::*;
 use failure;
 use proc_macro2::{Ident, Span, TokenStream};
 use query::QueryContext;
-use selection::{Selection, SelectionItem};
+use selection::{Selection, SelectionFragmentSpread, SelectionItem};
 use std::cell::Cell;
 use std::collections::BTreeSet;
 
@@ -46,40 +46,49 @@ pub(crate) fn union_variants(
                 true
             }
         }).map(|item| {
-            match item {
+            let (on, fields) = match item {
                 SelectionItem::Field(_) => Err(format_err!("field selection on union"))?,
-                SelectionItem::FragmentSpread(_) => Err(format_err!("fragment spread on union"))?,
-                SelectionItem::InlineFragment(frag) => {
-                    let variant_name = Ident::new(&frag.on, Span::call_site());
-                    used_variants.push(frag.on.to_string());
+                SelectionItem::FragmentSpread(SelectionFragmentSpread { fragment_name }) => {
+                    let fragment = query_context
+                        .fragments
+                        .get(fragment_name)
+                        .ok_or_else(|| format_err!("Unknown fragment: {}", &fragment_name))?;
 
-                    let new_prefix = format!("{}On{}", prefix, frag.on);
-
-                    let variant_type = Ident::new(&new_prefix, Span::call_site());
-
-                    let field_object_type = query_context.schema.objects.get(&frag.on).map(|_f| {
-                        query_context.maybe_expand_field(&frag.on, &frag.fields, &new_prefix)
-                    });
-                    let field_interface = query_context.schema.interfaces.get(&frag.on).map(|_f| {
-                        query_context.maybe_expand_field(&frag.on, &frag.fields, &new_prefix)
-                    });
-                    // nested unions, is that even a thing?
-                    let field_union_type = query_context.schema.unions.get(&frag.on).map(|_f| {
-                        query_context.maybe_expand_field(&frag.on, &frag.fields, &new_prefix)
-                    });
-
-                    match field_object_type.or(field_interface).or(field_union_type) {
-                        Some(tokens) => children_definitions.push(tokens?),
-                        None => Err(UnionError::UnknownType {
-                            ty: frag.on.to_string(),
-                        })?,
-                    };
-
-                    Ok(quote! {
-                        #variant_name(#variant_type)
-                    })
+                    (&fragment.on, &fragment.selection)
                 }
-            }
+                SelectionItem::InlineFragment(frag) => (&frag.on, &frag.fields),
+            };
+            let variant_name = Ident::new(&on, Span::call_site());
+            used_variants.push(on.to_string());
+
+            let new_prefix = format!("{}On{}", prefix, on);
+
+            let variant_type = Ident::new(&new_prefix, Span::call_site());
+
+            let field_object_type = query_context
+                .schema
+                .objects
+                .get(on)
+                .map(|_f| query_context.maybe_expand_field(&on, &fields, &new_prefix));
+            let field_interface = query_context
+                .schema
+                .interfaces
+                .get(on)
+                .map(|_f| query_context.maybe_expand_field(&on, &fields, &new_prefix));
+            let field_union_type = query_context
+                .schema
+                .unions
+                .get(on)
+                .map(|_f| query_context.maybe_expand_field(&on, &fields, &new_prefix));
+
+            match field_object_type.or(field_interface).or(field_union_type) {
+                Some(tokens) => children_definitions.push(tokens?),
+                None => Err(UnionError::UnknownType { ty: on.to_string() })?,
+            };
+
+            Ok(quote! {
+                #variant_name(#variant_type)
+            })
         }).collect();
 
     let variants = variants?;