+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## 0.6.0 (2018-12-30)
+
 ### Added
 
 - The CLI can now optionally format the generated code with rustfmt (enable the `rustfmt` feature).